사업자번호 글자수 이슈
long 처리

+++ finalEclips/src/main/java/com/example/finalEclips/eclips/mypage/dto/StoreInfoDto.java
@@ -12,7 +12,7 @@ public class StoreInfoDto {
     private char priceApproval;
     private char cleanlinessApproval;
     private String finalApprovalStatus;
-    private int businessRegNo;
+    private long businessRegNo;
     private String storeName;
     private String address;
     private String sidoName;
